Added Spell Tester, Fixed Evade Spells, Fixed No Disable Of Spells, Tweeked Some Settings

+++ EzEvade/EzEvade/Data/PositionInfo.cs
@@ -103,8 +103,8 @@ namespace EzEvade.Data
 
                 var spellDangerLevel = spell.Dangerlevel;
 
-                posDangerLevel = Math.Max(posDangerLevel, spellDangerLevel);
-                posDangerCount += spellDangerLevel;
+                posDangerLevel = Math.Max(posDangerLevel, (int) spellDangerLevel);
+                posDangerCount += (int)spellDangerLevel;
             }
 
             return new PositionInfo(
